Refactoring of MinimizationPsiManager and PsiProcessor and small docs
